+++ packages/nextjs/src/middleware.ts
@@ -4,6 +4,8 @@ import {
   NextResponse,
 } from 'next/server'
 
+import { getCss } from '@deepdish/workbench/css'
+
 export type DeepdishMiddlewareConfig = {
   draft: boolean
   verify: (request: NextRequest) => boolean | Promise<boolean>
@@ -22,6 +24,15 @@ export function deepdishMiddleware(
     const { url, method } = request
 
     if (url.includes('/__deepdish/') && method === 'GET') {
+      if (url.endsWith('/css')) {
+        const css = getCss()
+
+        const response = new NextResponse(css)
+        response.headers.set('Content-Type', 'text/css')
+
+        return response
+      }
+
       if (url.endsWith('/sign-in')) {
         return config.signIn(request)
       }