+++ packages/workbench/src/workbench.tsx
@@ -1,5 +1,0 @@
-import { Client } from './client'
-
-export async function Workbench() {
-  return <Client />
-}